#261: Add OAI-PMH ListRecords response support
The server will attempt to split the ListRecords response into multiple documents when validating. The URI of the split documents is `filename#oai-document-id`.

This commit also fixes a bug in ValidationView.recognizeDdiDocument() where an InputStreamResource would be fully consumed before testing to see if a valid document had been supplied.

This commit updates `cmv-core` to version 4.1.0-SNAPSHOT.